getting-started-with-geospatial-apps/case1-postgis-django/src/requirements.txt
celery==5.3.5
Django==4.2.7
django-browser-reload==1.12.0
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-opensearch-dsl==0.5.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
djangorestframework-simplejwt==5.3.0
djangorestframework-gis
drf-api-tracking==1.8.4
gunicorn==20.1.0
JSON-log-formatter==0.5.1
openapi==1.1.0
psycopg2-binary==2.9.9
redis==4.4.0
